backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.35
pydantic==2.9.0
pydantic-settings==2.5.0
python-jose[cryptography]==3.3.0
passlib==1.7.4
httpx==0.27.0
python-multipart==0.0.9
alembic==1.13.0
python-dotenv==1.0.1
